Big change
Change how data is read and written.
Add more attribute to account, user and book class.
Add labels that display intro and quantity of a book when the item in the book table is clicked.
Add 2 file "TempUsers.xml" and "TempAccounts.xml", they are used to temporarily store registration information so that admin or librarian can add it to the real database.
The way admin and librarian can accept an registration will be added in the next commit.