rewrite log in and log out feature
Now, when a user log in, depend on which role the user's account has, 1 of 3 windows (admin, librarian, reader) will be shown up.
A user can log out by simply click the "X" button on the upper right corner to close the window, the main window will reappear and the user can now log in with a nother account.